# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Joy VoyagerApi module adds REST Api end points to Voyager with Passport and Swagger support https://github.com/rxcod9/joy-voyager-api.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://user.uz",
            "description": "Primary API Server"
        },
        {
            "url": "http://youdo.test",
            "description": "Another API Server"
        }
    ],
    "paths": {
...
        "/api/same-tasks/{task}": {
            "get": {
                "tags": [
                    "TaskAPI"
                ],
                "summary": "Same tasks by Task ID",
                "operationId": "120ccd09ca7a2be2153ee7a0f2655447",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/responses/{task}": {
            "get": {
                "tags": [
                    "TaskAPI"
                ],
                "summary": "Response tasks",
                "operationId": "e40467954638499de8484738d2359d0c",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/task/{task}/response": {
            "post": {
                "tags": [
                    "TaskAPI"
                ],
                "summary": "Send Response",
                "operationId": "72b3712b916c380767bb13875dd42339",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/tasks-filter": {
            "get": {
                "tags": [
                    "TaskAPI"
                ],
                "summary": "Task filter",
                "operationId": "01811c7639caa6a28b67d031f8c8be43",
                "parameters": [
                    {
                        "name": "categories",
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "name": "lat",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "long",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "budget",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_remote",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "without_response",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "difference",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/task/{task}": {
            "get": {
                "tags": [
                    "TaskAPI"
                ],
                "summary": "Get Task By ID",
                "operationId": "f81261f5febbe5e9edecf62e8b165e8f",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/my-tasks-count": {
            "get": {
                "tags": [
                    "TaskAPI"
                ],
                "summary": "Get My Tasks Count",
                "operationId": "72718dc806703c197699e0f060cb0935",
                "parameters": [
                    {
                        "name": "is_performer",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/my-tasks": {
            "get": {
                "tags": [
                    "TaskAPI"
                ],
                "summary": "Get My Tasks",
                "operationId": "1a716f0c34b249580b5d4194baa5f168",
                "parameters": [
                    {
                        "name": "is_performer",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/create-task/routing": {
            "post": {
                "tags": [
                    "TaskAPI"
                ],
                "summary": "Routing",
                "operationId": "75bde0adcddefe8478e20c8b7daeb051",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/for_del_new_task/{task}": {
            "delete": {
                "tags": [
                    "Search"
                ],
                "summary": "Delete Task",
                "operationId": "944c123df58af87508b1e98eda756420",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "description": "Bearer token",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}